posts work correctly with logged in user

+++ src/main/java/com/example/demo/infrastructure/Post/PostService.java
@@ -3,8 +3,9 @@ package com.example.demo.infrastructure.Post;
 import com.example.demo.domain.entities.PostEntity;
 import com.example.demo.domain.entities.UserEntity;
 import com.example.demo.domain.repositories.PostRepository;
-import com.example.demo.domain.repositories.UserRepository;
+import com.example.demo.security.UserVerificationProvider;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,11 +16,10 @@ import java.util.Optional;
 public class PostService {
 
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
 
     public PostDto createPost(PostDto postDto) {
-        UserEntity user = userRepository.findById(postDto.getUserId()).orElseThrow();
-        PostEntity post = PostDto.mapToEntity(postDto, user);
+        UserEntity userEntity = (UserEntity) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        PostEntity post = PostDto.mapToEntity(postDto, userEntity);
         postRepository.save(post);
         return postDto;
     }
@@ -33,20 +33,20 @@ public class PostService {
         return new PostDto(post);
     }
 
-//    public List<PostDto> getPostsOfUser(){
-//
-//    }
-
     public List<PostDto> getPostsByCategory(String category) {
         return postRepository.getPostsByCategory(category).stream().map(PostDto::new).toList();
     }
 
     public String deletePost(Long id) {
+        PostEntity postEntity = postRepository.findById(id).orElseThrow();
+        UserVerificationProvider.checkPostAuthor(postEntity);
         postRepository.deleteById(id);
         return "Post which id is " + id + " was deleted succesfully";
     }
 
     public Optional<PostDto> editPost(PostDto postDto, Long id) {
+        PostEntity postEntity = postRepository.findById(id).orElseThrow();
+        UserVerificationProvider.checkPostAuthor(postEntity);
         return postRepository.findById(id).map((post -> {
             post.setTitle(postDto.getTitle());
             post.setCategory(postDto.getCategory());
